fix: Fix usage of wrong HTTP verb to delete files
FileController used a get mapping to handle users' requests to delete a
file. This was the wrong method for an action that is not idempotent and
it was forced upon the backend by the usage in the frontend of an anchor
element to implement the delete buttons, which do not allow http methods
other than GET. The anchor element in the home template has now been
replaced with a button element wrapped inside a form.